Rename todo.rs -> pending_contributions.rs
