+++ handel/src/aggregation.rs
@@ -17,9 +17,9 @@ use crate::{
     level::Level,
     network::{LevelUpdateSender, Network},
     partitioner::Partitioner,
+    pending_contributions::{TodoItem, TodoList},
     protocol::Protocol,
     store::ContributionStore,
-    todo::{TodoItem, TodoList},
     update::LevelUpdate,
     verifier::{VerificationResult, Verifier},
     Identifier,
